backend/requirements.txt
# Web Framework
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
motor==3.6.0
beanie==1.29.0
redis==5.2.1

# AI/ML
openai==1.61.0
faiss-cpu==1.8.0
numpy==1.26.4
httpx==0.28.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.0.1
cryptography==44.0.1

# Data Validation
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

# Utilities
gunicorn==23.0.0